feat: [DP-372] remove dependecy on common package for setup.py
