+++ setup.py
@@ -6,7 +6,27 @@ import sys
 from setuptools import setup, find_packages
 from setuptools.command.test import test as test_command
 
-from hip_data_tools.common import get_release_version, get_long_description
+
+def get_release_version():
+    """
+    Gets the Release version based on the latest git tag from GIT_TAG env var, else returns 0.0
+    Returns: string containing version for the release
+
+    """
+    git_version = os.getenv("GIT_TAG", "v0.0")
+    pypi_version = git_version.lstrip("v").strip()
+    return pypi_version
+
+
+def get_long_description():
+    """
+    Get the contents of readme file as long_description
+    Returns: bytes containing readme file
+
+    """
+    file_path = os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'README.md'))
+    with open(file_path) as readme_file:
+        return readme_file.read()
 
 
 class PyTest(test_command):
